sort

+++ src/main/java/com/example/Paginationdemo/repository/StudentRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepo extends JpaRepository<StudentDepartment,Long>, JpaSpecificationExecutor<StudentDepartment> {
+public interface StudentRepo extends JpaRepository<StudentDepartment,Integer>, JpaSpecificationExecutor<StudentDepartment> {
 }
